refacto supression des photo lie aux mandat et properties

+++ src/database/drizzle/repositories/photos.ts
@@ -3,6 +3,7 @@ import { photos } from "../schema/photos";
 import { eq, sql } from "drizzle-orm";
 import { extractIdFromUrl, removeFile, UPLOAD_DIR_PROPERTIES } from "@/lib/fileSystem";
 import path from "path";
+import { UPLOAD_DIR_IMAGES } from "@/config/dir";
 
 export const insertPhoto = async (data: any) => {
     const prepare = db
@@ -55,17 +56,20 @@ export const deletePhotoByID = async (id: number | string) => {
     }
 };
 
+type contextPathType = "properties" | "perquisitions";
+
 /**
  * Suppression des photos + fichiers
  * @param ids list des id des photos
+ * @param contextPath le dossier contenant les fichiers
  */
-export const removePhotosByAndFile = async (ids: number[] | string[]) => {
+export const removePhotosByAndFile = async (ids: number[] | string[], contextPath: contextPathType) => {
     if (ids) {
         for (const id of ids) {
             const photo = await getOnePhotosByID(id);
             if (photo) {
                 const key = extractIdFromUrl(photo.url);
-                remove(key);
+                remove(key, contextPath);
                 await deletePhotoByID(photo.id);
             }
         }
@@ -74,10 +78,13 @@ export const removePhotosByAndFile = async (ids: number[] | string[]) => {
 
 /**
  * Suppression du fichier lié aux photos
+ * @param key identifiant unique utilisé aux renommage de fichier + extension
+ * @param folder le dossier contenant les fichiers
  */
-export const remove = async (key: string) => {
+export const remove = async (key: string, folder: string) => {
     try {
-        const filepath = path.join(UPLOAD_DIR_PROPERTIES, key);
+        if (!folder) throw new Error("filepath is missing");
+        const filepath = path.join(UPLOAD_DIR_IMAGES, folder, key);
         await removeFile(filepath);
     } catch (error: any) {
         throw new Error(error.message);
